+++ src/Regressors/Adaline.php
@@ -14,7 +14,7 @@ use Rubix\ML\NeuralNet\FeedForward;
 use Rubix\ML\Other\Traits\LoggerAware;
 use Rubix\ML\NeuralNet\Optimizers\Adam;
 use Rubix\ML\NeuralNet\Layers\Continuous;
-use Rubix\ML\NeuralNet\Layers\Placeholder;
+use Rubix\ML\NeuralNet\Layers\Placeholder1D;
 use Rubix\ML\NeuralNet\Optimizers\Optimizer;
 use Rubix\ML\NeuralNet\CostFunctions\CostFunction;
 use Rubix\ML\NeuralNet\CostFunctions\LeastSquares;
@@ -110,8 +110,8 @@ class Adaline implements Online, Verbose, Persistable
      * @throws \InvalidArgumentException
      * @return void
      */
-    public function __construct(int $batchSize = 50, ?Optimizer $optimizer = null, float $alpha = 1e-4,
-                            int $epochs = 1000, float $minChange = 1e-4, ?CostFunction $costFn = null)
+    public function __construct(int $batchSize = 100, ?Optimizer $optimizer = null, float $alpha = 1e-4,
+                                int $epochs = 1000, float $minChange = 1e-4, ?CostFunction $costFn = null)
     {
         if ($batchSize < 1) {
             throw new InvalidArgumentException('Cannot have less than 1 sample'
@@ -194,7 +194,7 @@ class Adaline implements Online, Verbose, Persistable
         }
 
         $this->network = new FeedForward(
-            new Placeholder($dataset->numColumns()),
+            new Placeholder1D($dataset->numColumns()),
             [],
             new Continuous($this->alpha, $this->costFn),
             $this->optimizer